157 resolve project dependency problems
