+++ dao/src/main/java/greencity/repository/EventRepo.java
@@ -1,25 +1,25 @@
 package greencity.repository;
 
-import greencity.dto.event.*;
+import greencity.dto.event.model.EventModelDto;
 import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
 
-@Repository
-public class EventRepo {
+//@Repository
+//public class EventRepo {
 
-    private List<EventModelDto> events = new ArrayList<>();
-    private AtomicLong eventId = new AtomicLong();
+    //private List<EventModelDto> events = new ArrayList<>();
+    //private AtomicLong eventId = new AtomicLong();
 
-    public EventModelDto save(EventModelDto event) {
-        event.setId(eventId.incrementAndGet());
-        events.add(event);
+    //public EventModelDto save(EventModelDto event) {
+        //event.setId(eventId.incrementAndGet());
+      //  events.add(event);
 
-        return event;
-    }
+    //    return event;
+  //  }
 
 
 
-}
+//}
