+++ service-api/src/main/java/greencity/dto/event/model/EventImage.java
@@ -1,4 +1,4 @@
-package greencity.dto.event;
+package greencity.dto.event.model;
 
 import lombok.*;
