+++ service-api/src/main/java/greencity/dto/event/model/EventLocationAddress.java
@@ -1,4 +1,4 @@
-package greencity.dto.event;
+package greencity.dto.event.model;
 
 import lombok.Builder;
 import lombok.Getter;
