+++ service-api/src/main/java/greencity/dto/event/model/EventLocationLink.java
@@ -1,9 +1,8 @@
-package greencity.dto.event;
+package greencity.dto.event.model;
 
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
-import lombok.experimental.SuperBuilder;
 
 @Getter
 @Setter
