+++ service-api/src/main/java/greencity/service/EventService.java
@@ -1,7 +1,7 @@
 package greencity.service;
 
 import greencity.dto.event.EventRequestSaveDto;
-import greencity.dto.event.EventModelDto;
+import greencity.dto.event.model.EventModelDto;
 import greencity.dto.user.UserVO;
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.multipart.MultipartFile;
